+++ apps/hardware_benchmarks/tests/counter/counter_generator.cpp
@@ -6,8 +6,8 @@ using namespace Halide;
 
 class UnitTestCounter : public Halide::Generator<UnitTestCounter> {
 public:
-    Input<Buffer<int16_t>>  input{"input", 2};
-    Output<Buffer<int16_t>> output{"output", 2};
+    Input<Buffer<uint8_t>>  input{"input", 2};
+    Output<Buffer<uint8_t>> output{"output", 2};
 
     void generate() {
         /* THE ALGORITHM */
@@ -22,7 +22,7 @@ public:
         county(x,y) = y + hw_input(x,y);
 
         Func hw_output("hw_output");
-        hw_output(x, y) = cast<int16_t>(countx(x, y) + county(x, y));
+        hw_output(x, y) = cast<uint8_t>(countx(x, y) + county(x, y));
         output(x, y) = hw_output(x,y);
 
         /* THE SCHEDULE */
